Fix indeterminacy and use generators.
First, there was a bug in ``.nextmeeting CHANNEL`` where it would arbitrarily
return different meetings as the next one up.  Now, this explicit sorts the
meetings it gets back so we know that the next one is really the next one.

Second, this makes use of generators and ``itertools`` to avoid flattening the
meetings into a list (for fun!)