More flexible karma detection.
This will allow it so that the USERNAME++ bit can appear *anywhere* in the line
that the user sends to irc.

Furthermore, we publish the entire original line in the fedmsg message so that
we can potentially record *why* people got karma if someone says 'thanks for
all the fish threebean++'.