Adjust the .mirroradmins command for MirrorManager2
From the next release (0.2.0) of MirrorManager2, there will be a new API
endpoint dedicated to providing admins of the specified host or site.

This change will not work until this new release is done, but that will
not change anything from the current situation where it does not work
either.