fix caching
previously, we used list_all_entities which was slow, and caused
the stg fasjson to give a 500, since list_users() gives us everything
we need, we now use that instead.

Also, add some checks and backups if the IRC nicks list is None. which
is the case if a user has not defined an IRC nick

Signed-off-by: Ryan Lerch <rlerch@redhat.com>