Remove hardcoded calendar URLs
This removes the hardcoded calendar URLs, so we can in theory set this
up for ursabot on staging if we really want to.

freenode was also in here, and the calendars were updated to the new
char service, so this was not working becasue of that also. So changed
this too

Signed-off-by: Ryan Lerch <rlerch@redhat.com>